refactor(data): moved type definitions in a separate file
