+++ module/data/__init__.py
@@ -1,3 +1,4 @@
 """Data module"""
-from .constants import START_TEXT, CLEAR_LOGFILE_TEXT
-from .config import config_map, PageConfig, GroupConfig, Config
+from .constants import START_TEXT, CLEAR_LOGFILE_TEXT, DEFAULT_NOTICES_DATA
+from .config import config_map
+from .types import Config, GroupConfig, NoticeData, PageConfig
