+++ module/data/config.py
@@ -1,37 +1,7 @@
 """Configuration map"""
 import os
-from typing import TypedDict
 import yaml
-
-
-class PageConfig(TypedDict):
-    """page sub-configuration type definition"""
-
-    label: str
-    urls: "list[str]"
-    channels: "list[str | int]"
-
-
-class GroupConfig(TypedDict):
-    """group sub-configuraiton type definition"""
-
-    base_url: str
-    approval_group_chatid: "int | None"
-    pages: "dict[str, PageConfig]"
-
-
-class Config(TypedDict):
-    """config_map type definition"""
-
-    token: str
-    dev_group_chatid: int
-    log_group_chatid: int
-    logfile_reset_interval_minutes: int
-    update_interval: int
-    max_messages_length: int
-    max_length_footer: str
-    max_connection_tries: int
-    notices_groups: "dict[str, GroupConfig]"
+from .types import Config
 
 
 def load_configurations(path: str = "config/") -> Config:
